requirements.txt
Pillow==11.0.0
rich==13.9.4
